# services/product-service/docs/api-contract.yaml
openapi: 3.0.3
info:
  title: Product Service API
  description: >
    API del catálogo de categorías, productos base localizados y productos de
    usuario monolingües para ShoppingList. Los productos base y de usuario se
    exponen como colecciones independientes y sus DTOs reflejan sus reglas de
    ciclo de vida, internacionalización y propiedad.


    Nota de vigencia: este documento es el contrato de diseño de Fase 1,
    redactado antes de implementar el servicio. Una vez exista código y
    springdoc-openapi genere el contrato real a partir de él, este fichero se
    marcará como snapshot histórico de diseño (no se mantendrá sincronizado a
    mano), y el contrato vigente pasará a ser el expuesto en /v3/api-docs.
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Entorno local (Docker Compose)
...
paths:
  /categories:
    get:
      tags: [categories]
      summary: Listar categorías del catálogo
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Página de categorías, traducidas al idioma solicitado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryPage'
    post:
      tags: [categories]
      summary: Crear una categoría del catálogo con sus traducciones
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Categoría creada.
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Ya existe una categoría con ese code.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /categories/{id}:
    get:
      tags: [categories]
      summary: Obtener una categoría por id
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Categoría encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          $ref: '#/components/responses/NotFound'
  /base-products:
    get:
      tags: [base-products]
      summary: Listar productos base
      description: >
        Listado público del catálogo de productos base. Devuelve nombre y
        descripción localizados según `Accept-Language` con el fallback de
        ADR-011.
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
            nullable: true
        - name: text
          in: query
          description: Búsqueda textual en el nombre localizado.
          schema:
            type: string
            nullable: true
      responses:
        '200':
          description: Página de productos base con nombre y descripción localizados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseProductPage'
    post:
      tags: [base-products]
      summary: Crear un producto base
      description: >
        Operación administrativa. Mismo criterio placeholder que ADR-006 para
        propietario: sin auth real hasta Fase 4
        (ADR-XXF4-keycloak-como-identity-provider pendiente).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBaseProductRequest'
      responses:
        '201':
          description: Producto base creado.
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseProduct'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Ya existe un producto base con ese code.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /base-products/{id}:
    get:
      tags: [base-products]
      summary: Obtener un producto base por id
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Producto base encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseProduct'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [base-products]
      summary: Editar un producto base
      description: >
        Operación administrativa. Mismo criterio placeholder que ADR-006 para
        propietario: sin auth real hasta Fase 4
        (ADR-XXF4-keycloak-como-identity-provider pendiente).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBaseProductRequest'
      responses:
        '200':
          description: Producto base actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseProduct'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Ya existe otro producto base con ese code.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    delete:
      tags: [base-products]
      summary: Eliminar un producto base
      description: >
        Operación administrativa. Mismo criterio placeholder que ADR-006 para
        propietario: sin auth real hasta Fase 4
        (ADR-XXF4-keycloak-como-identity-provider pendiente).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Producto base eliminado.
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: >
        Idioma solicitado para entidades del catálogo (es | en | eu). Si no se
        especifica o el idioma no tiene traducción, se aplica fallback a `en`.
      schema:
        type: string
        enum: [es, en, eu]
        default: en
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    NotFound:
      description: El recurso solicitado no existe.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    ValidationError:
      description: >
        Error de validación de la petición. Los errores de Bean Validation
        responden con code VALIDATION_FAILED y el array errors por campo; los
        errores de negocio 400 (categoría inexistente, locale no soportado,
        etc.) usan el shape ProblemDetail sin errors.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblemDetail'
  schemas:
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: dairy
        name:
          type: string
          description: Traducción en el idioma resuelto (Accept-Language, con fallback a en).
        isActive:
          type: boolean
    CategoryPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
    CreateCategoryRequest:
      type: object
      required: [code, isActive, translations]
      properties:
        code:
          type: string
          minLength: 1
          example: dairy
        isActive:
          type: boolean
        translations:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Translation'
    Translation:
      type: object
      required: [locale, name]
      properties:
        locale:
          type: string
          minLength: 2
          maxLength: 5
          example: es
        name:
          type: string
          maxLength: 128
          example: Lácteos
    BaseProduct:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        categoryId:
          type: string
          format: uuid
        defaultUnit:
          $ref: '#/components/schemas/UnitEnum'
        calories:
          type: integer
          format: int32
          nullable: true
        caloriesPer:
          $ref: '#/components/schemas/CaloriesPerEnum'
        isActive:
          type: boolean
        name:
          type: string
          description: Nombre resuelto mediante i18n y fallback.
        description:
          type: string
          nullable: true
          description: Descripción resuelta mediante i18n y fallback.
...
    BaseProductPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BaseProduct'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
...
    CreateBaseProductRequest:
      type: object
      required: [code, categoryId, defaultUnit, caloriesPer, isActive, translations]
      properties:
        code:
          type: string
          minLength: 1
        categoryId:
          type: string
          format: uuid
        defaultUnit:
          $ref: '#/components/schemas/UnitEnum'
        calories:
          type: integer
          format: int32
          nullable: true
        caloriesPer:
          $ref: '#/components/schemas/CaloriesPerEnum'
        isActive:
          type: boolean
        translations:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ProductTranslation'
    UpdateBaseProductRequest:
      type: object
      properties:
        code:
          type: string
        categoryId:
          type: string
          format: uuid
        defaultUnit:
          $ref: '#/components/schemas/UnitEnum'
        calories:
          type: integer
          format: int32
          nullable: true
        caloriesPer:
          $ref: '#/components/schemas/CaloriesPerEnum'
        isActive:
          type: boolean
        translations:
          type: array
          description: Reemplaza el conjunto completo de traducciones.
          items:
            $ref: '#/components/schemas/ProductTranslation'
...
    ProductTranslation:
      type: object
      required: [locale, name]
      properties:
        locale:
          type: string
          minLength: 2
          maxLength: 5
        name:
          type: string
          maxLength: 128
        description:
          type: string
          nullable: true
    UnitEnum:
      type: string
      enum: [KG, G, L, ML, UNIT, DOZEN]
    CaloriesPerEnum:
      type: string
      enum: [G, ML, UNIT]
...
    ProblemDetail:
      type: object
      description: >
        Respuesta de error según RFC 9457 (application/problem+json) con la
        extensión code del contrato. El catálogo completo de códigos vive en
        el enum ErrorCode (ver ADR-014); las excepciones no controladas
        responden 500 con code INTERNAL_ERROR, fuera del catálogo.
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
          description: Resumen estable del error en inglés, el mismo para todas las ocurrencias.
        status:
          type: integer
          format: int32
        detail:
          type: string
          description: Mensaje específico en inglés; puede incluir el identificador del recurso implicado.
        instance:
          type: string
          format: uri
          nullable: true
          description: URI del recurso de la petición que produjo el error.
        code:
          type: string
          enum: [CATEGORY_NOT_FOUND, INVALID_CATEGORY, BASE_PRODUCT_NOT_FOUND, USER_PRODUCT_NOT_FOUND, PRODUCT_NOT_FOUND, DUPLICATE_PRODUCT_CODE, DUPLICATE_CATEGORY_CODE, UNSUPPORTED_LOCALE, NAME_REQUIRED, DEFAULT_UNIT_REQUIRED, CALORIES_PER_REQUIRED, OWNER_MISMATCH, INVALID_PRODUCT_TYPE, INVALID_BASE_PRODUCT, VALIDATION_FAILED]
          description: Identificador machine-readable del error; el contrato entre el backend y el frontend.
    FieldError:
      type: object
      description: Error de validación de un campo concreto de la petición.
      properties:
        code:
          type: string
          description: Nombre de la constraint que falla (FieldError.getCode(), p.ej. NotBlank).
        field:
          type: string
          description: Nombre del campo que falla.
        message:
          type: string
          description: Mensaje de la constraint en inglés (FieldError.getDefaultMessage()).
    ValidationProblemDetail:
      type: object
      allOf:
        - $ref: '#/components/schemas/ProblemDetail'
        - type: object
          properties:
            errors:
              type: array
              description: Errores por campo; presente solo en errores de validación Bean Validation.
              items:
                $ref: '#/components/schemas/FieldError'