# services/product-service/src/main/java/dev/ivborrezo/shoppinglist/product/service/category/entity/Category.java
package dev.ivborrezo.shoppinglist.product.service.category.entity;

import dev.ivborrezo.shoppinglist.product.service.common.UuidV7;
import jakarta.persistence.CascadeType;
import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.EntityListeners;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.OneToMany;
import jakarta.persistence.PrePersist;
import jakarta.persistence.Table;
import java.time.Instant;
import java.util.LinkedHashSet;
import java.util.Set;
import java.util.UUID;
import org.jspecify.annotations.Nullable;
import org.springframework.data.annotation.CreatedDate;
import org.springframework.data.annotation.LastModifiedDate;
import org.springframework.data.jpa.domain.support.AuditingEntityListener;

/**
 * Categoría del catálogo gestionado por el sistema.
 *
 * <p>Los nombres localizados viven en {@code category_translation} (patrón i18n Table),
 * relacionados por la colección {@code translations}. Las columnas {@code created_at}/{@code
 * updated_at} se almacenan como {@code TIMESTAMPTZ} en PostgreSQL y se mapean a {@link Instant}
 * (instante absoluto, sin zona adjunta); la convención es transversal a todas las columnas de
 * auditoría del monorepo.
 */
@Entity
@Table(name = "category")
@EntityListeners(AuditingEntityListener.class)
public class Category {

  @Id
  @GeneratedValue(strategy = GenerationType.IDENTITY)
  private Long id;

  @Column(name = "public_id", nullable = false, updatable = false)
  private @Nullable UUID publicId;

  @Column(nullable = false, length = 64)
  private String code;

  @Column(nullable = false)
  private Boolean isActive;

  @CreatedDate
  @Column(nullable = false, updatable = false)
  private Instant createdAt;

  @LastModifiedDate
  @Column(nullable = false)
  private Instant updatedAt;

  @OneToMany(mappedBy = "category", cascade = CascadeType.ALL, orphanRemoval = true)
  private Set<CategoryTranslation> translations = new LinkedHashSet<>();

  public Category() {}

  @PrePersist
  void assignPublicId() {
    if (publicId == null) {
      publicId = UuidV7.generate();
    }
  }

  public @Nullable Long getId() {
    return id;
  }

  public void setId(Long id) {
    this.id = id;
  }

  public @Nullable UUID getPublicId() {
    return publicId;
  }

  public void setPublicId(UUID publicId) {
    this.publicId = publicId;
  }

  public String getCode() {
    return code;
  }

  public void setCode(String code) {
    this.code = code;
  }

  public Boolean getIsActive() {
    return isActive;
  }

  public void setIsActive(Boolean isActive) {
    this.isActive = isActive;
  }

  public Instant getCreatedAt() {
    return createdAt;
  }

  public Instant getUpdatedAt() {
    return updatedAt;
  }

  public Set<CategoryTranslation> getTranslations() {
    return translations;
  }
}